+++ migrate.php
@@ -1567,6 +1567,7 @@ $newSettings = [
     'news_per_page'           => '10',
     'blog_per_page'           => '10',
     'events_per_page'         => '10',
+    'blog_authors_index_enabled' => '0',
     'comments_enabled'        => '1',
     'comment_moderation_mode' => 'always',
     'comment_close_days'      => '0',